requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Templating
jinja2==3.1.3

# AI/ML
google-generativeai==0.4.0
scikit-learn==1.4.0
numpy==1.26.3

# PDF Processing
pdfplumber==0.10.3

# NLP
nltk==3.8.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Utilities
tenacity==8.2.3

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.23.3
httpx==0.26.0
